fix(daytona): restore behaviors lost in the SDK migration
Audit of the migration commit (41ffbc01) found several regressions on top
of the pointer-vs-value bug. This commit restores the missing behaviors:

- CreateSandbox now refuses an empty SnapshotID instead of falling back to
  the bogus "hiveloop/bridge:latest" string the old code carried around.
  Production never hit this path (orchestrator.resolveSnapshot always
  returns a snapshot id), but the silent fallback would have masked
  config bugs.
- Restore the Name field on SandboxBaseParams — sandboxes were being
  created with auto-generated names instead of "hiveloop-ded-...".
- Restore the defensive copy of opts.EnvVars; the SDK keeps a reference
  and the caller's later mutations would leak through.
- Restore typed not-found detection via errors.As against
  *daytonaerrors.DaytonaNotFoundError instead of substring matching the
  message text.
- Restore the 120s timeout on ExecuteCommand. The old hand-rolled client
  hardcoded it; the SDK defers to caller context, and cloneAgentRepositories
  runs git clones with no deadline.
- Populate SnapshotStatusResult.ErrorMsg from the SDK's collapsed
  ErrorReason so orchestrator_template.go's fallback message logic still
  reports the real error instead of "snapshot build failed with unknown
  error".
